// frontend/src/app/api/generated/models/practice-material-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Flashcard } from '../models/flashcard';
import { PracticeQuestion } from '../models/practice-question';

/**
 * Response containing generated practice materials for an exam upload.
 */
export interface PracticeMaterialResponse {
  course_id: number;
  flashcards: Array<Flashcard>;
  generated_at: string;
  id: number;
  questions: Array<PracticeQuestion>;
  student_pid: number;
  upload_id: number;
  weak_topics: Array<string>;
}